upscale/requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0

# Video processing
imageio>=2.21.0
imageio-ffmpeg>=0.4.7
ffmpeg-python>=0.2.0

# Real-ESRGAN and upscaling
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.8
realesrgan>=0.2.5

# SSH utilities
paramiko>=2.11.0
scp>=0.14.0

# Web server
flask>=2.0.0

# Additional utilities
requests>=2.25.0
wget>=3.2
PyYAML>=6.0

# System monitoring
psutil>=5.9.0